Remove `alias_method_chain`
